moved extensions class to separate file
